update core api

+++ core/src/main/java/io/ffreedom/redstone/core/position/api/PositionProducer.java
@@ -4,9 +4,9 @@ package io.ffreedom.redstone.core.position.api;
 public interface PositionProducer<T extends Position> {
 
 	default T produce(int accountId, int instrumentId) {
-		return produce(accountId, instrumentId, 0);
+		return produce(accountId, instrumentId, 0L);
 	}
 
-	T produce(int accountId, int instrumentId, double qty);
+	T produce(int accountId, int instrumentId, long qty);
 
 }
